refactor: split utils.py into different util files according to logic
