+++ src/run.py
@@ -11,12 +11,16 @@ from MetricHandler import metric_request_from_prompt, MetricHandler
 from ReportReader import read_sast_report_html
 from Utils.embedding_utils import generate_code_embeddings
 from Utils.utils import (
-    read_cve_html_file,
     print_conclusion,
-    get_human_verified_results,
-    validate_configurations,
-    load_config,
-    download_repo, print_config
+    validate_environment
+)
+from Utils.file_utils import (
+    create_embeddings_for_all_project_files, 
+    read_known_errors_file,
+    get_human_verified_results
+)
+from Utils.html_utils import (
+    read_cve_html_file, 
 )
 from model.EvaluationSummary import EvaluationSummary
 from model.SummaryInfo import SummaryInfo
